renamed component file names, reordered directories and moved out of utils folder into more standard ./lib folder and updated imports
