+++ src/components/CodeDetailsLogoComponent.tsx
