+++ src/components/Projects/ProjectComponents/ProjectFormComponent.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useProjects } from "@/providers/projects-provider";
 import { useSupabaseToken } from "@/hooks/use-SupabaseClerkJWTToken";
-import { AddProjectForm } from "@/components/Projects/AddProjectForm/page";
+import { AddProjectForm } from "@/components/Projects/ProjectComponents/AddProjectFormComponent";
 
 export function ProjectForm() {
   const { token } = useSupabaseToken();
