+++ src/lib/FormattedDate.tsx
@@ -1,5 +1,5 @@
 // Create this in src/components/ui/FormattedDate.tsx
-import { ClientOnly } from "@/utils/ClientSideUtils";
+import { ClientOnly } from "@/lib/ClientSideUtils";
 import React from "react";
 
 export function FormattedDate({
